requirements.txt
av==11.0.0
celluloid==0.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
easing-functions==1.0.4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib_resources==6.1.3
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.1
lazy_loader==0.3
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
nibabel==5.2.1
nilearn==0.10.3
nltools==0.5.0
numexpr==2.8.4
numpy==1.23.5
packaging==23.2
pandas==2.0.3
pillow==10.2.0
py-feat==0.6.1
pynv==0.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
sympy==1.12
threadpoolctl==3.3.0
tifffile==2023.7.10
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xgboost==2.0.3
zipp==3.17.0
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.15
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0